Move move refactoring to its own class and file
