+++ src/main/kotlin/org/purescript/psi/declaration/MoveValueDeclarationRefactoring.kt
@@ -1,0 +1,32 @@
+package org.purescript.psi.declaration
+
+import com.intellij.psi.search.GlobalSearchScope
+import com.intellij.psi.search.searches.ReferencesSearch
+import com.intellij.refactoring.BaseRefactoringProcessor
+import com.intellij.refactoring.move.MoveHandler
+import com.intellij.usageView.BaseUsageViewDescriptor
+import com.intellij.usageView.UsageInfo
+import org.purescript.psi.module.Module
+
+class MoveValueDeclarationRefactoring(
+    private val toMove: PSValueDeclaration,
+    private val targetModule: Module.Psi
+) : BaseRefactoringProcessor(toMove.project) {
+    override fun createUsageViewDescriptor(usages: Array<out UsageInfo>) =
+        BaseUsageViewDescriptor(toMove)
+
+    override fun findUsages(): Array<UsageInfo> =
+        ReferencesSearch
+            .search(toMove, GlobalSearchScope.projectScope(toMove.project))
+            .findAll()
+            .map(::UsageInfo)
+            .toTypedArray()
+
+    override fun performRefactoring(usages: Array<out UsageInfo>) {
+        val first = (toMove.signature?: toMove).prevSibling // whitespace
+        targetModule.addRange(first, toMove)
+        toMove.module?.deleteChildRange(first, toMove)
+    }
+
+    override fun getCommandName(): String = MoveHandler.getRefactoringName()
+}
