+++ src/main/kotlin/org/purescript/psi/declaration/ValueDeclarationMoveHandlerDelegate.kt
@@ -7,16 +7,11 @@ import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiReference
 import com.intellij.psi.search.GlobalSearchScope
-import com.intellij.psi.search.searches.ReferencesSearch
-import com.intellij.refactoring.BaseRefactoringProcessor
 import com.intellij.refactoring.move.MoveHandler
 import com.intellij.refactoring.move.MoveHandlerDelegate
 import com.intellij.refactoring.ui.RefactoringDialog
 import com.intellij.ui.dsl.builder.bindItem
 import com.intellij.ui.dsl.builder.panel
-import com.intellij.usageView.BaseUsageViewDescriptor
-import com.intellij.usageView.UsageInfo
-import com.intellij.usageView.UsageViewDescriptor
 import org.purescript.PSLanguage
 import org.purescript.psi.module.Module
 import org.purescript.psi.module.ModuleNameIndex
@@ -77,31 +72,9 @@ class ValueDeclarationMoveHandlerDelegate : MoveHandlerDelegate() {
             )
             val targetModule = modules
                 .single { ".spago" !in it.containingFile.virtualFile.path }
-            invokeRefactoring(object : BaseRefactoringProcessor(project) {
-                override fun createUsageViewDescriptor(usages: Array<out UsageInfo>): UsageViewDescriptor {
-                    return BaseUsageViewDescriptor(element)
-                }
-
-                override fun findUsages(): Array<UsageInfo> =
-                    ReferencesSearch
-                        .search(
-                            element,
-                            GlobalSearchScope.projectScope(project)
-                        )
-                        .findAll()
-                        .map(::UsageInfo)
-                        .toTypedArray()
-
-                override fun performRefactoring(usages: Array<out UsageInfo>) {
-                    val first = (element.signature?: element).prevSibling // whitespace
-                    targetModule.addRange(first, element)
-                    element.module?.deleteChildRange(first, element)
-                }
-
-                override fun getCommandName(): String =
-                    MoveHandler.getRefactoringName()
-
-            })
+            invokeRefactoring(
+                MoveValueDeclarationRefactoring(element, targetModule)
+            )
         }
 
         init {
@@ -112,3 +85,4 @@ class ValueDeclarationMoveHandlerDelegate : MoveHandlerDelegate() {
         }
     }
 }
+
